// ticket-management-api/src/entities/enums/ticket.enums.ts
export enum TicketTopic {
  BILLING = 'billing',
  BUG = 'bug',
  FEATURE = 'feature',
  OTHER = 'other',
}

export enum TicketPriority {
  LOW = 'low',
  MEDIUM = 'medium',
  HIGH = 'high',
}

export enum TicketStatus {
  CREATED = 'created',
  IN_PROGRESS = 'in_progress',
  COMPLETED = 'completed',
}